requirements.txt
jinja2
piexif
pillow
